requirements.txt
gunicorn==20.0.4
flask==1.1.2
flask_restx==0.2.0
flask_CORS==3.0.9
SQLAlchemy==1.3.20
flask_sqlalchemy==2.4.4
Werkzeug==0.16.1
requests==2.24.0
psycopg2-binary==2.8.6
alembic==1.5.8
python-jose[cryptography]
google-cloud-pubsub==2.5.0
sqlalchemy_filters